+++ week7/src/components/Write/WriteForm.jsx
@@ -2,30 +2,7 @@ import styled from "styled-components";
 import { useNavigate } from "react-router-dom";
 import { client } from "../../lib/api";
 import { useRef } from "react";
-
-const postInfoList = [
-  {
-    label: "이름",
-    id: "name",
-    placeholder: "이름이 뭐예요?",
-  },
-  {
-    label: "내용",
-    id: "content",
-    placeholder: "무슨 내용의 편지를 써볼까요?",
-  },
-  {
-    label: "비밀번호",
-    id: "password",
-    placeholder: "비밀번호를 통해 편지를 잠궈요.",
-    type: "password",
-  },
-  {
-    label: "비밀번호 힌트",
-    id: "hint",
-    placeholder: "비밀번호 힌트를 주세요.",
-  },
-];
+import Inputs from "../common/Inputs";
 
 function WriteForm() {
   const navigate = useNavigate();
@@ -35,9 +12,6 @@ function WriteForm() {
     e.preventDefault();
 
     const formData = new FormData(e.target);
-    for (var pair of formData.entries()) {
-      console.log(pair[0] + ", " + pair[1]);
-    }
     await client.post("/letter", formData);
     navigate("/");
   };
@@ -45,17 +19,7 @@ function WriteForm() {
   return (
     <WriteFormWrapper>
       <PostForm onSubmit={handleSubmit}>
-        {postInfoList.map(({ label, id, placeholder, type }) => (
-          <InputWrapper key={id}>
-            <InputLabel htmlFor={id}>{label}</InputLabel>
-            <input
-              type={type || "text"}
-              placeholder={placeholder}
-              id={id}
-              name={id}
-            />
-          </InputWrapper>
-        ))}
+        <Inputs />
         <InputLabel htmlFor="images">썸네일</InputLabel>
         <FileUploadButton
           onClick={(e) => {
@@ -96,33 +60,6 @@ const PostForm = styled.form`
   display: flex;
   flex-direction: column;
   align-items: center;
-`;
-
-const InputWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  label {
-    padding: 8px;
-
-    font-size: 25px;
-    font-weight: bold;
-    font-family: ${({ theme }) => theme.font.title};
-  }
-
-  input {
-    width: 350px;
-    height: auto;
-
-    margin-bottom: 20px;
-    border-radius: 10px;
-    border: none;
-
-    text-align: center;
-    font-size: 18px;
-    font-family: ${({ theme }) => theme.font.content};
-  }
 `;
 
 const InputLabel = styled.label`
